client/react: avoid duplicate requests caused by undefined user/tenant
inlining useAccounts because it seemed more trouble than it's worth
